+++ pyinaturalist/controllers/projects.py
@@ -39,3 +39,5 @@ class ProjectController(BaseController):
             )
             responses.append(response)
         return responses
+
+    # TODO: Add update_project()